Introduce zypper.py for zypper related logic
